Move deploy related functions from cli.rs to deploy module in deploy.rs, and re-export moved functions
